Replace most assert() calls with require()
As per docs assert should be only used for stuff that indicates a bug
(if result is ex thrown). So asserts are only used for as defensive
double-checking tool. Everything else like that's expected to happen,
like validating external inputs, most of the stuff in modifiers, etc
should use require:
http://solidity.readthedocs.io/en/develop/control-structures.html?highlight=require#error-handling-assert-require-revert-and-exceptions

It also mentions this (require throws revert-based exception):
"Note that assert-style exceptions consume all gas available to the call,
 while revert-style exceptions will not consume any gas starting from the
 Metropolis release."