Use number of blocks for round duration
During testing round duration turned out to be confusing and unreliable,
partially due to how blocks_per_hour was calculated (starting with 60/17
division which results in 3 in solidity). Also, it was impossible to
test shorter duration than one hour.

Therefore, amend the contract to use duration in blocks as an input
argument. This moves concept of time from contract to deploy script,
simplifying the contract and eliminating unnecessary calculations in
solidity.